Phase 5: Complete K12 LMS implementation with UI improvements and bug fixes
- Fixed icon sizes across student dashboard and related pages
- Resolved API endpoint 404 errors (double /api/api/ issue)
- Enhanced UI with gradient backgrounds and modern styling
- Improved accessibility with focus traps and ARIA attributes
- Added comprehensive error handling and loading states
- Implemented design tokens and semantic color system
- Added CSV export functionality for gradebook
- Enhanced seed data with multiple classes and assignments
- Added logging middleware and rate limiting
- Implemented demo mode with floating tips
- Added comprehensive documentation and Docker setup
- Fixed component import/export issues
- Cleaned up project structure by removing summary files

All Phase 5 requirements completed successfully.

+++ backend/app/services/recommendations.py
@@ -56,7 +56,7 @@ def compute_skill_mastery(student_id: int, db: Session) -> Dict[str, float]:
     skill_mastery = {}
     for skill_tag, scores in skill_scores.items():
         if scores:
-            skill_mastery[skill_tag] = sum(scores) / len(scores)
+            skill_mastery[skill_tag] = float(sum(scores) / len(scores))
         else:
             skill_mastery[skill_tag] = 0.0
     
@@ -204,9 +204,9 @@ def rank_lessons_for_student(student_id: int, class_id: int, db: Session, k: int
             'lesson_id': lesson.id,
             'title': lesson.title,
             'reason': reason,
-            'score': combined_score,
+            'score': float(combined_score),
             'skill_tags': lesson.skill_tags,
-            'skill_mastery': {tag: skill_mastery.get(tag, 0.0) for tag in lesson.skill_tags}
+            'skill_mastery': {tag: float(skill_mastery.get(tag, 0.0)) for tag in lesson.skill_tags}
         })
     
     # Sort by score (highest first) and return top k
